requirements.txt
Flask
mysql-connector-python
azure-storage-blob
twilio
python-dateutil
python-dotenv